Move newtork classes into a package in utils
