+++ app/src/androidTest/java/ch/epfl/sweng/SDP/utils/NetworkStateReceiverTest.java
@@ -4,6 +4,8 @@ import android.support.test.InstrumentationRegistry;
 
 import org.junit.Test;
 
+import ch.epfl.sweng.SDP.utils.network.NetworkStateReceiver;
+
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
