+++ app/src/main/java/ch/epfl/sweng/SDP/utils/network/NetworkStateListenerWrapper.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP.utils;
+package ch.epfl.sweng.SDP.utils.network;
 
 import android.app.Dialog;
 import android.content.Context;
