+++ app/src/main/java/ch/epfl/sweng/SDP/utils/network/NetworkStateReceiver.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP.utils;
+package ch.epfl.sweng.SDP.utils.network;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
