+++ app/src/main/java/ch/epfl/sweng/SDP/utils/network/NetworkStateReceiverListener.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP.utils;
+package ch.epfl.sweng.SDP.utils.network;
 
 public interface NetworkStateReceiverListener {
     void networkAvailable();
